==> 01_create_wikipedia-based_tokenizer.ipynb <==
import sqlite3
import pandas as pd
import gc

### Convert Wikipedia sqlite database to a pandas dataframe

db_path = "data/wikipedia/docs.db"
connection = sqlite3.connect(db_path, check_same_thread=False)
cursor = connection.cursor()

cursor.execute("SELECT id, text FROM documents")
data_json = {"id": [], "text": []}
for r in cursor.fetchall():
    data_json["id"].append(r[0]); data_json["text"].append(r[1])
cursor.close()
connection.close()

data_df = pd.DataFrame.from_dict(data_json)

del data_json
gc.collect()

# data_df.to_csv("datasets_lisandro/wikipedia.csv", index=False)
data_df.head(5)

### Create Wikipedia-based tokenizer and save

from sklearn.feature_extraction.text import TfidfVectorizer
import pickle

vectorizer = TfidfVectorizer(
                            strip_accents='unicode',
                            analyzer='word',
                            stop_words='english',
                            lowercase=True,
                            ngram_range=(1,2))
                            
vectorizer.fit(data_df["text"])
pickle.dump(vectorizer, open("models_lisandro/2gram_tfidf.pkl", "wb"))
